Renamed proto for linting, added functions to package lint
